some tweaks to make it work on 64bit archs

+++ pspdis/src/bfd_replacement.c
@@ -13,30 +13,31 @@
 #               to satisfy libopcodes.
 #
 */
+#include "util.h"
 #include "bfd_replacement.h"
 
 bfd_vma
 bfd_getl32 (const void *p)
 {
   const bfd_byte *addr = p;
-  unsigned long v;
+  u32 v;
 
-  v = (unsigned long) addr[0];
-  v |= (unsigned long) addr[1] << 8;
-  v |= (unsigned long) addr[2] << 16;
-  v |= (unsigned long) addr[3] << 24;
+  v = (u32) addr[0];
+  v |= (u32) addr[1] << 8;
+  v |= (u32) addr[2] << 16;
+  v |= (u32) addr[3] << 24;
   return v;
 }
 bfd_vma
 bfd_getb32 (const void *p)
 {
   const bfd_byte *addr = p;
-  unsigned long v;
+  u32 v;
 
-  v = (unsigned long) addr[0] << 24;
-  v |= (unsigned long) addr[1] << 16;
-  v |= (unsigned long) addr[2] << 8;
-  v |= (unsigned long) addr[3];
+  v = (u32) addr[0] << 24;
+  v |= (u32) addr[1] << 16;
+  v |= (u32) addr[2] << 8;
+  v |= (u32) addr[3];
   return v;
 }
 bfd_vma
